add a test to test the test system
I'm in a bit of a rush, so I didn't try running the testfiles
from the main build.lua file; rather added a new one just for
tests. N.B. it takes regression-test.tex from l3build/ not
from ../support/

I'm using %&pdftex to run the test file in plain;
is that the right idea?