Typeset in dedicated location
This approach avoids as much as is sensible the danger of picking up
'stray' files. Notice that the typesetting directory is not currently cleaned
out at the start of 'doc': potentially we want to allow all PDFs to be here
and copy them out later in the process than is currently the case. At the
moment, each PDF is copied back to the 'parent' location at the end of the
run: should this be the case?