Use DocStrip version of regression-test, etc.
This removes the need to have a copy of regression-test.tex in support.
At the same time, as the 'standard' position is that nothing is run in isolation
it does not make sense to copy any support files. Instead, do this just for
our own code, and be selective about waht we copy (presumably the best plan).