Remove -output-directory from cmdcheck target (see #213)
There was also an issue here in that this function does seem to need
path searching in all cases. For the present, this is enable generally
but that probably needs revising.

Note that a quick check shows up some issues here to do with the
data recorded by l3doc: will log separately.