Normalise long lines
Lines which hit exactly max_print_lines can be detected and recombined.
Life is made a bit tricky by LuaTeX: it's 'out by one' in most cases.
However, if the line is broken with a space dropped from the result then
LuaTeX isn't out, so we end up needing a couple of new .tlg files to
allow for that, although we drop more.

An update for LaTeX2e to match this change will be required:
coming up in a bit.