Use argparse-like syntax
This checkin switches to an argparse-liek syntax for setting the
engine. As a result, we can now have multiple test names for check/save
targets. The same option now applies to the format target for l3kernel.

A new -H option is added to halt tests at first failure.

This fixes #245 and #246.

Changes may be needed to any customised build.lua scripts, most
obviously that for LaTeX2e.