Improve normalisation for chars 0/127/28-255
For ^^@ and ^^?, LuaTeX needs some work to get into line with
other engines.

For the top-half of the 8-bit range, for pdfTeX/XeTeX/LuaTeX
it makes sense to normalise to ^^<n> format, done using .tcx
file for pdfTeX and a search for UTF-8 chars for XeTeX/LuaTeX.
(This will be useful for LaTeX2e testing.) For (u)pTeX we seem
to be out of luck.