Drop use of SVN revision
The SVN revision is awkward as a version string as it can only
be known by carefully co-ordinating checkin to the master repo.
Moreover, it's easy to be 'out-by-one' and/or to find that
last-minute fixes require multiple commits.

Here, we assume that the date alone will be a suitable identifier
for releases. In particular, the development versions will at some
stage (soon) be marked up by l3build when installing locally, so
there should be no issue there.