replace relative with absolute paths
The relpath function is broken in multiple places. A comment directs it
to look from the source, followed directly by a line looking from the
target. In the component comparison routine, a "target" also takes the
place of a "source".  These bugs never surfaced because the paths fed to
the functions always were of equal length.

However, fixing them just reveals the deeper flaws of the algorithm. It
cannot deal with the current directory (".") correctly, which is fed to
it in many places.

Due to the nature of Lua, a sane, portable implementation is quite
complex. Instead, ripping it out and relying solemnly on absolute paths
in its place proves to be more easy and robust. Inspiration for this
change came from Waf [1], a general-purpose build system, which also
largely refrains from explicit relative paths altogether.

Note that absolute paths are required by tools such as MakeIndex when
allowing typesetting in complex set ups (for example when not
everything is in the working directory).

[1]: https://waf.io/