allow for subdirectory globbing
The new function tree wraps filelist, extending it to process not only
files in the given directory, but also the whole subdirectory hierarchy.
Its name is derived from the command-line tool "tree", which is to "ls"
what the tree function is to filelist.

It could serve as a drop-in replacement for filelist in all places, but
is only carefully applied as other functions are not yet prepared to
deal with paths in the place of simple file names. For now, it is only
employed with copying, enabling the user to collect files from different
location to be combined for unpacking, typesetting or testing.