Use a single table for all options
This gives us one point of contact for adding/modifying options.

We could perhaps make argparse() available more widely based on this
model, with the list and perhaps command line data passed as arguments.

At present only 1-1 mappings are supported but that is easy to
change: simply requires one or two additional keys.