omit leading ./ from output path
Explicitly anchoring the output path at the local directory "." gives
off the impression this path should be understood relative to the
current working directory build.lua was executed from. While the
directory structure inside the build directory mostly resembles the
source layout, this is not always the case, especially when dealing with
submodules.

Apart from all that, the leading dot does not add any merit to the
pathname expressiveness.