resolve copied links
When any file copied into a build directory is a symbolic link, the link
could not correctly point to its target from its new location, should
the target not coincidentally be copied as well. Errors of the kind
"file not found" will then occur for the copied links.  This is true for
targets inside as well as outside of the project, as long as they are
specified relatively.

Resolving the links, copying the file content of the target instead of
the link, eliminates this problem.

As the documentation of the xcopy utility [1], specifically the "/b"
option tells, resolving links was the default behaviour in Windows all
along.

[1] https://technet.microsoft.com/de-de/library/cc771254(v=ws.10).aspx