Normalise out box numbers from \show<box>
With a29a0363056df5aa6f9ac796228f8fc0776fd967, this will remove most
register-based variation in .tlg files. Left are explicit cases, for
example showing the \meaning of a register (boxes show as \char so could
not easily be handled there in any case).

As with the parent commit, this requires new .tlg files in some cases.