Replace literal "." by currentdir
This is going to be useful as it splits the cases where "." means
'current dir' from other use cases. Use where "." is used in e.g. path
splitting are retained: they are in a way distinct to the current dir
concept.

There are a few places that the split between 'current dir' and 'main
dir' is not clear, as well as those that need further attention to have
a fully-working docfiledir and to introduce sourcefildir. Those may need
further work/discussion.