Better install --dry-run
Key point here is that the 'cleanup' step is omitted.

Formatting of data is also improved.

We may wish to make the cleanup step less agressive once an uninstall
target is available: one might remove only the files matching those to
be created. (Or is that a bad idea? Aligns with what uninstall should
do in the first place.)